+++ src/main/java/blackjack/domain/Card.java
@@ -1,8 +1,5 @@
 package blackjack.domain;
 
-import blackjack.util.CardNumber;
-import blackjack.util.CardSuit;
-
 public class Card {
 
     private final CardNumber cardNumber;
