+++ src/main/java/blackjack/domain/CardNumber.java
@@ -1,4 +1,4 @@
-package blackjack.util;
+package blackjack.domain;
 
 import java.util.stream.Stream;
