+++ src/main/java/blackjack/domain/Cards.java
@@ -1,8 +1,7 @@
 package blackjack.domain;
 
-import blackjack.util.CardNumber;
 import blackjack.util.CardPickerGenerator;
-import blackjack.util.CardSuit;
+
 import java.util.ArrayList;
 import java.util.List;
