+++ src/test/java/blackjack/domain/CardsTest.java
@@ -2,9 +2,7 @@ package blackjack.domain;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-import blackjack.util.CardNumber;
 import blackjack.util.CardPickerGenerator;
-import blackjack.util.CardSuit;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
