+++ src/test/java/blackjack/domain/PlayerTest.java
@@ -3,8 +3,6 @@ package blackjack.domain;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import blackjack.util.CardNumber;
-import blackjack.util.CardSuit;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
